Made rubocop support not required
ref #3

Before if the code base did not has the rubocop report file the
analyzing process had been failed. This was a very unconvinient way to
provide an additional ability for the end user. This commit changes this
behaviour and of there is no report file the system skips rubocop
sensor.