Feature/Rubocop (#10)
* Added Rubocop rules inside SQ profile file

ref #3

* Added rubocop rules definitions repo XML

ref #3

* Added project property for Rubocop report path

ref #3

* Added Rubocop report data models

ref #3

* Added loading cops rules definitions

ref #3

* Implemented rubocop report json parser

ref #3

* Completed RubocopSensor

ref #3

* Register RubocopSensor and report json parser

ref #3

* Fixed rule status naming in rule definitions

ref #3

Actually all rule statuses should be uppercased to be conformed with
enumerable org.sonar.api.rule.RuleStatus

* Fixed rubocop report parser

- Added required annotations
- Fixed parsing `corrected` property of an offense - in some cases it is
  null
- Updated readme with Rubocop mentions

ref #3

* Made rubocop support not required

ref #3

Before if the code base did not has the rubocop report file the
analyzing process had been failed. This was a very unconvinient way to
provide an additional ability for the end user. This commit changes this
behaviour and of there is no report file the system skips rubocop
sensor.

* Rebased the rubocop PR

* removed and optimized imports